+++ module/Application/src/LazyQuahogClientFactory.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Application;
+
+use Interop\Container\ContainerInterface;
+use Zend\ServiceManager\Factory\FactoryInterface;
+
+class LazyQuahogClientFactory implements FactoryInterface
+{
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    {
+        return new $requestedName($container->get('config'));
+    }
+}