+++ sql/rollback_1.3.0_1.4.0.sql
@@ -2,11 +2,11 @@
 #--
 #-- System      : SenDb
 #--
-#-- Filename    : rollback_1.2.0_1.3.0.sql
+#-- Filename    : rollback_1.3.0_1.4.0.sql
 #--
 #-- Description : Migration script
 #--
-#--               1.2.0    <--    1.3.0
+#--               1.3.0    <--    1.4.0
 #--
 #-----------------------------------------------------------------------------------------------------------------------
 
@@ -42,7 +42,7 @@ migrate:BEGIN
         versionTableExists;
 
     IF (versionTableExists = 0) THEN
-        SELECT 'This script can only run on a version 1.3.0 database' AS result
+        SELECT 'This script can only run on a version 1.4.0 database' AS result
         UNION
         SELECT 'If you are certain you want to run this script, modify the source to bypass this check'
         UNION
@@ -61,8 +61,8 @@ migrate:BEGIN
     INTO
         currentVersion;
 
-    IF (currentVersion <> '1.3.0') THEN
-        SELECT 'This script can only run on a version 1.3.0 database' AS result
+    IF (currentVersion <> '1.4.0') THEN
+        SELECT 'This script can only run on a version 1.4.0 database' AS result
         UNION
         SELECT 'If you are certain you want to run this script, modify the source to bypass this check'
         UNION
@@ -74,23 +74,19 @@ migrate:BEGIN
                                                             #-----------------------------------------------------------
                                                             #-- Version is correct - migrate
                                                             #-----------------------------------------------------------
-    DROP TABLE IF EXISTS users;
+    #-- Not feasible to rollback the changes to default values or the MyISAM to InnoDB migration.
+    DROP TABLE IF EXISTS event_attachment;
 
                                                             #-----------------------------------------------------------
                                                             #-- Log deployment
                                                             #-----------------------------------------------------------
-    INSERT INTO version (version) VALUES ('1.2.0');
+    INSERT INTO version (version) VALUES ('1.3.0');
 
-    SELECT 'Rollback complete. Now at version 1.2.0.' AS result;
+    SELECT 'Rollback complete. Now at version 1.3.0.' AS result;
 
 END //
 
 DELIMITER ;
-
-#-- CREATE / DROP TRIGGER statements are not permitted in a stored procedure, so they must be done separately.
-DROP TRIGGER IF EXISTS maintainUsersAfterInsertOnScagroup;
-DROP TRIGGER IF EXISTS maintainUsersAfterUpdateOnScagroup;
-DROP TRIGGER IF EXISTS maintainUsersAfterDeleteOnScagroup;
 
 CALL spMigrate();